+++ exonum-java-binding/core/rust/integration_tests/tests/service_runtime.rs
@@ -59,9 +59,9 @@ fn bootstrap() {
         },
     );
 
-    let mut testkit = TestKitBuilder::validator()
-        .with_service(service_runtime.service_proxy())
-        .create();
+    let service = service_runtime.create_service("", TEST_SERVICE_MODULE_NAME);
+
+    let mut testkit = TestKitBuilder::validator().with_service(service).create();
 
     testkit.create_block();
 }